% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mean_sd_trends.R
\name{calculate_mean_sd_trends}
\alias{calculate_mean_sd_trends}
\title{Calculate the Mean-Variance trend}
\usage{
calculate_mean_sd_trends(data, design_matrix)
}
\arguments{
\item{data}{A \code{tibble} or \code{data.frame} to annotate with mean and sd}

\item{design_matrix}{A design matrix for the data (see example)}
}
\value{
A \code{tibble} or \code{data.frame} with the mean and sd vectors
}
\description{
Calculates the mean and sd of the rows.
}
\examples{
# Please see the vignette for a tutorial
# vignette('baldur-tutorial')
}
