% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_priors.R
\name{estimate_gamma_priors}
\alias{estimate_gamma_priors}
\title{Estimate Gamma priors for sigma}
\usage{
estimate_gamma_priors(data, design_matrix, gamma_reg)
}
\arguments{
\item{data}{A \code{tibble} or \code{data.frame} to add gamma priors to}

\item{design_matrix}{A design matrix for the data (see example)}

\item{gamma_reg}{A Gamma regression object}
}
\value{
A \code{tibble} or \code{data.frame} with the alpha,beta priors
}
\description{
Estimates the priors for the Bayesian model.
}
\examples{
# Setup model matrix
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))
colnames(design) <- paste0("ng", c(50, 100))

# Normalize data
yeast_norm <- yeast \%>\%
    psrn("identifier") \%>\%
    # Get mean-variance trends
    calculate_mean_sd_trends(design)
# Fit gamma regression
gam_reg <- fit_gamma_regression(yeast_norm, sd ~ mean)

# Estimate priors
yeast_norm \%>\%
    estimate_gamma_priors(design, gam_reg)
}
