% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgmr_plotting.R
\name{lgmr_plotting}
\alias{lgmr_plotting}
\alias{plot_lgrm_regression}
\alias{plot_regression_field}
\title{Visualization of LGMR models}
\usage{
plot_lgrm_regression(model)

plot_regression_field(model, n = 10, rng = 10)
}
\arguments{
\item{model}{A LGMR model object}

\item{n}{Number of interpolation points for each peptides regression}

\item{rng}{The proportional range of each peptides regression.
E.g., a value of 10 will make each line span 1 \% of the x-axis.}
}
\value{
A ggplot object
}
\description{
Visualization of LGMR models
}
