% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gamma_regression.R
\name{plot_gamma_partition}
\alias{plot_gamma_partition}
\title{Function for plotting the gamma regression after partitioning}
\usage{
plot_gamma_partition(data, design, ...)
}
\arguments{
\item{data}{The data to use for producing the plots.}

\item{design}{A design matrix as produced by \code{stats::\link[stats:model.matrix]{model.matrix}}.}

\item{...}{Additional arguments to \code{baldur::\link[baldur:trend_partitioning]{trend_partitioning}}.}
}
\value{
a plot with the mean-variance trend after partitioning
}
\description{
Generates a scatter plot with the gamma regressions of the mean-variance
trends after partitioning.
}
\examples{
# Produce a design matrix
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))
colnames(design) <- paste0("ng", c(50, 100))

# Normalize and log transform the data
yeast <- psrn(yeast, "identifier")

# Generate the plot
\dontrun{
yeast \%>\%
  calculate_mean_sd_trends(design) \%>\%
  plot_gamma_partition(design, verbose = FALSE)
}
}
