% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gamma_regression.R
\name{plot_gamma_regression}
\alias{plot_gamma_regression}
\title{Function for plotting the mean-variance gamma regressions}
\usage{
plot_gamma_regression(data, design, ...)
}
\arguments{
\item{data}{The data to use for producing the plots.}

\item{design}{A design matrix as produced by \code{\link[stats]{model.matrix}}.}

\item{...}{Additional arguments to \code{\link[baldur]{trend_partitioning}}.}
}
\value{
a plot with the mean-variance trend before partitioning on the left side, and the right side after.
}
\description{
Generates a scatter plot with the gamma regressions of the mean-variance
trends for the precision weights and imputation.
}
\examples{
# Produce a design matrix
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))
colnames(design) <- paste0("ng", c(50, 100))

# Normalize and log transform the data
yeast <- psrn(yeast_prog, "identifier")

# Generate the plots
\dontrun{
plot_gamma_regression(yeast, design, verbose = FALSE)
}
}
