% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_posterior.R
\name{sample_posterior}
\alias{sample_posterior}
\title{Sample the Posterior of the Bayesian model}
\usage{
sample_posterior(
  data,
  id_col_name,
  design_matrix,
  contrast_matrix,
  uncertainty_matrix,
  bayesian_model = stanmodels$uncertainty_model,
  clusters = 1,
  robust = F,
  perc = 0.05,
  ...
)
}
\arguments{
\item{data}{A \code{tibble} or \code{data.frame} with alpha,beta priors annotated}

\item{id_col_name}{A character of the id column}

\item{design_matrix}{A design matrix for the data (see example)}

\item{contrast_matrix}{A contrast matrix of the decisions to test such that the first column is the index of the column in the design matrix that should be compared to the second column}

\item{uncertainty_matrix}{A matrix of observation uncertainty}

\item{bayesian_model}{Which Bayesian model to use. Currently only one internal model allowed, this argument is for forward compatibility}

\item{clusters}{The number of parallel threads to run on.}

\item{robust}{If integration of the posterior should be done robust by trimming the tails}

\item{perc}{If robust is \code{TRUE} how much of each tail to remove before integration. Should be between 0 (no trimming) and 0.5 (trimming all data).}

\item{...}{Additional arguments passed to}
}
\value{
A \code{tibble} or \code{data.frame} annotated  with statistics of the posterior
}
\description{
Function to sample the posterior of the Bayesian decision model.
}
\examples{
# (Please see the vignette for a tutorial)
# Setup model matrix
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))
colnames(design) <- paste0("ng", c(50, 100))

yeast_norm <- yeast \%>\%
# Remove missing data
  tidyr::drop_na() \%>\%
  # Normalize data
  psrn('identifier') \%>\%
  # Add mean-variance trends
  calculate_mean_sd_trends(design)
# Fit the gamma regression
gam <- fit_gamma_regression(yeast_norm, sd ~ mean)
# Estimate each data point's uncertainty
unc <- estimate_uncertainty(yeast_norm, 'identifier', design, gam)
yeast_norm <- yeast_norm \%>\%
   # Add hyper-priors for sigma
   estimate_gamma_priors(design, gam)
# Setup contrast matrix
contrast <- matrix(1:2, ncol = 2)
\dontrun{
yeast_norm \%>\%
  sample_posterior(
    'identifier',
    design,
    contrast,
    unc,
    clusters = 1 # I highly recommend increasing the number of parallel workers/clusters
                 # this will greatly reduce the speed of running baldur
  )
}
}
