% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_partitioning.R
\name{trend_partitioning}
\alias{trend_partitioning}
\title{Mean-Variance Trend Partitioning}
\usage{
trend_partitioning(
  data,
  design_matrix,
  formula = sd ~ mean + c,
  eps = 0.1,
  n = 1000,
  verbose = T
)
}
\arguments{
\item{data}{A \code{tibble} or \code{data.frame} to partition}

\item{design_matrix}{A design matrix for the data (see example)}

\item{formula}{Formula for the Gamma regression}

\item{eps}{Size of the integration window}

\item{n}{Number of integration windows}

\item{verbose}{If the number of points moved should be output}
}
\value{
A \code{tibble} or \code{data.frame} the partitioning vector \code{c}
}
\description{
Partitions the data into two mixtures.
}
\examples{
# Please see the vignette for a tutorial
# vignette('baldur-tutorial')
}
