% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weakly_informative.R
\name{weakly_informative}
\alias{weakly_informative}
\title{Doc for weakly_informative
... // however many lines of doc
...}
\description{
Doc for weakly_informative
... // however many lines of doc
...
}
