% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baldur-package.R
\docType{package}
\name{baldur-package}
\alias{baldur-package}
\alias{baldur}
\title{The 'baldur' package.}
\description{
baldur is a Bayesian hierarchical model for statistical decision in proteomics data
}
\references{
Stan Development Team (2022). RStan: the R interface to Stan. R package version 2.21.5. https://mc-stan.org
}
