% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baldur3-package.R
\docType{package}
\name{baldur3-package}
\alias{baldur3-package}
\alias{baldur3}
\title{The 'baldur3' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2022). RStan: the R interface to Stan. R package version 2.21.5. https://mc-stan.org
}
