% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_priors.R
\name{estimate_gamma_priors}
\alias{estimate_gamma_priors}
\title{Estimate Gamma priors for sigma}
\usage{
estimate_gamma_priors(data, design_matrix, formula = sd ~ mean + c)
}
\arguments{
\item{data}{A \code{tibble} or \code{data.frame} to add gamma priors to}

\item{design_matrix}{A design matrix for the data (see example)}

\item{formula}{Formula for the Gamma regression}
}
\value{
A \code{tibble} or \code{data.frame} with the alpha,beta priors
}
\description{
Estimates the priors for the Bayesian model.
}
\examples{
'lorem'
}
