% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_posterior.R
\name{sample_posterior}
\alias{sample_posterior}
\title{Sample Posterior}
\usage{
sample_posterior(
  data,
  id_col_name,
  design_matrix,
  contrast_matrix,
  uncertainty_matrix,
  bayesian_model = stanmodels$uncertainty_model,
  clusters = 1
)
}
\arguments{
\item{data}{A \code{tibble} or \code{data.frame} with alpha,beta priors annotated}

\item{id_col_name}{A character of the id column}

\item{design_matrix}{A design matrix for the data (see example)}

\item{contrast_matrix}{A contrast matrix of the decisions to test such that the first column is the index of the column in the design matrix that should be compared to the second column}

\item{uncertainty_matrix}{A matrix of observation uncertainty}

\item{bayesian_model}{Which Bayesian model to use. Currently only on internal model allowed, this argument is for forward compatibility}

\item{clusters}{The number of parallel threads to run on.}
}
\value{
A \code{tibble} or \code{data.frame} annotated  with statistics of the posterior
}
\description{
Function to sample the posterior of the Bayesian decision model.
}
\examples{
'lorem'
}
