% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_posterior.R
\name{sample_posterior}
\alias{sample_posterior}
\title{Sample the Posterior of the Bayesian model}
\usage{
sample_posterior(
  data,
  id_col_name,
  design_matrix,
  contrast_matrix,
  uncertainty_matrix,
  bayesian_model = stanmodels$uncertainty_model,
  clusters = 1,
  robust = F,
  perc = 0.05,
  mu_not = 0,
  ...
)
}
\arguments{
\item{data}{A \code{tibble} or \code{data.frame} with alpha,beta priors annotated}

\item{id_col_name}{A character of the id column}

\item{design_matrix}{A design matrix for the data (see example)}

\item{contrast_matrix}{A contrast matrix of the decisions to test such that the first column is the index of the column in the design matrix that should be compared to the second column}

\item{uncertainty_matrix}{A matrix of observation uncertainty}

\item{bayesian_model}{Which Bayesian model to use. Currently only one internal model allowed, this argument is for forward compatibility}

\item{clusters}{The number of parallel threads to run on.}

\item{robust}{If integration of the posterior should be done robust by trimming the tails}

\item{perc}{If robust is \code{TRUE} how much of each tail to remove before integration. Should be between 0 (no trimming) and 0.5 (trimming all data).}

\item{mu_not}{The value of the null hypothesis for the difference in means}

\item{...}{Additional arguments passed to \code{rstan::\link[rstan:sampling]{sampling}}. Note that verbose will always be forced to \code{FALSE}}
}
\value{
A \code{tibble} or \code{data.frame} annotated  with statistics of the posterior and p(error).
\code{err} is the probability of error, i.e., the two tail-density supporting the null-hypothesis, \code{lfc} is the estimated log$_2$-fold change, \code{sigma} is the common #' variance, and \code{lp} is the log-posterior.
Columns without suffix shows the mean estimate from the posterior, while the suffixes \verb{_025}, \verb{_50}, and \verb{_975}, are the 2.5, 50.0, and 97.5, percentiles, respectively.
The suffixes \verb{_eff} and \verb{_rhat} are the diagnostic variables returned by \code{rstan} (please see the Stan manual for more details).
In general, a larger \verb{_eff} indicates a better sampling efficiency, and \verb{_rhat} compares the mixing within chains against between the chains and should be smaller than 1.05.
}
\description{
Function to sample the posterior of the Bayesian decision model.
}
\examples{
# (Please see the vignette for a tutorial)
# Setup model matrix
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))
colnames(design) <- paste0("ng", c(50, 100))

yeast_norm <- yeast \%>\%
# Remove missing data
  tidyr::drop_na() \%>\%
  # Normalize data
  psrn('identifier') \%>\%
  # Add mean-variance trends
  calculate_mean_sd_trends(design)
# Fit the gamma regression
gam <- fit_gamma_regression(yeast_norm, sd ~ mean)
# Estimate each data point's uncertainty
unc <- estimate_uncertainty(yeast_norm, 'identifier', design, gam)
yeast_norm <- yeast_norm \%>\%
   # Add hyper-priors for sigma
   estimate_gamma_priors(design, gam)
# Setup contrast matrix
contrast <- matrix(1:2, ncol = 2)
\dontrun{
yeast_norm \%>\%
  sample_posterior(
    'identifier',
    design,
    contrast,
    unc,
    clusters = 1 # I highly recommend increasing the number of parallel workers/clusters
                 # this will greatly reduce the speed of running baldur
  )
}
}
