% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_partitioning.R
\name{trend_partitioning}
\alias{trend_partitioning}
\title{Mean-Variance Trend Partitioning}
\usage{
trend_partitioning(
  data,
  design_matrix,
  formula = sd ~ mean + c,
  eps = 0.1,
  n = 1000,
  verbose = TRUE
)
}
\arguments{
\item{data}{A \code{tibble} or \code{data.frame} to partition}

\item{design_matrix}{A design matrix for the data (see example)}

\item{formula}{Formula for the Gamma regression}

\item{eps}{Size of the integration window}

\item{n}{Number of integration windows}

\item{verbose}{If the number of points moved should be output}
}
\value{
A \code{tibble} or \code{data.frame} the partitioning vector \code{c}
}
\description{
Partitions the data into two mixtures.
}
\examples{
# (Please see the vignettes for tutorials)
# Setup model matrix
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))
colnames(design) <- paste0("ng", c(50, 100))

yeast_norm <- yeast \%>\%
# Remove missing data
  tidyr::drop_na() \%>\%
  # Normalize data
  psrn('identifier') \%>\%
  # Add mean-variance trends
  calculate_mean_sd_trends(design)
\donttest{
yeast_norm \%>\%
  # Partition the data
  trend_partitioning(design)
}
}
