% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weakly_informative.R
\name{weakly_informative}
\alias{weakly_informative}
\title{Baldur's weakly informative prior for the mean in conditions}
\description{
Here we will model the mean of the prior with a weakly informative (WI) prior.
We will assume that, in essence, nothing is know about the mean.
Following from
}
\details{
\deqn{\boldsymbol{Y}\sim\mathcal{N}(\mu_0 + \boldsymbol{\eta}\sigma,\sigma\boldsymbol{u})}
}
